Move throttle check to the wrapper
This ensures it is done consistently for all our entities

It also gets us away from the horrible signature on the old
function.

However, I have to do some more testing to see how the
Exception behaves, not being an extension of Core_Exception

That makes sense in the run-time context but does the
scheduled job handle it well?